requirements.txt
streamlit
vertexai
pillow
pandas
reportlab
python-dateutil
google-cloud-aiplatform
